requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
bidict==0.23.1
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
distro==1.9.0
fastapi==0.135.2
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
llvmlite==0.46.0
numba==0.64.0
numpy==2.2.6
openai==2.30.0
passlib==1.7.4
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dotenv==1.2.2
python-engineio==4.13.1
python-socketio==5.16.1
requests==2.33.0
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==1.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
websockets==16.0
wsproto==1.3.2
